remove cache for initial PR
